move check_privilege.* to main
